+++ src/main/java/DisplayTree.java
@@ -80,25 +80,9 @@ public class DisplayTree {
 
 
     public static void main(String[] args) {
-        Tree tree = perfect(128);
+        Tree tree = Utils.perfect(128);
         String html = display(tree);
         System.out.println(html);
     }
 
-    public static Tree perfect(int size) {
-        Tree tree = new Tree();
-        perfectChild(tree, 0, size);
-        return tree;
-    }
-
-    public static void perfectChild(Tree t, int init, int end) {
-        int size = end - init;
-        if (size < 2) {
-            return;
-        }
-        int key = size / 2;
-        t.inserir(init+key);
-        perfectChild(t, init, init + key);
-        perfectChild(t, init + key , end);
-    }
 }
