+++ src/main/java/SameKeyException.java
@@ -1,5 +1,3 @@
-package src;
-
 public class SameKeyException extends RuntimeException {
     public SameKeyException(String msg) {
         super(msg);
